Feedback flows and request structure

+++ src/main/java/com/coffee/roastnbrew/resources/UserResource.java
@@ -2,9 +2,8 @@ package com.coffee.roastnbrew.resources;
 
 import com.coffee.roastnbrew.models.User;
 import com.coffee.roastnbrew.services.UserService;
-
 import com.coffee.roastnbrew.utils.RestUtils;
-import java.util.Map;
+
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import javax.ws.rs.*;
@@ -46,7 +45,7 @@ public class UserResource {
     
     @DELETE
     @Path("/{id}")
-    public Response deleteUser(@PathParam("id") int userId) {
+    public Response deleteUser(@PathParam("id") long userId) {
         return RestUtils.ok(userService.deleteUser(userId));
     }
 }
